requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.5.0
cryptography==42.0.0
dataclasses-json==0.6.3
distro==1.9.0
Django==4.2.9
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-compressor==2.4.1
django-compressor-toolkit==0.6.0
docopt==0.6.2
dukpy==0.2.3
et-xmlfile==1.1.0
faiss-cpu==1.7.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.2
langchain-community==0.0.14
langchain-core==0.1.14
langsmith==0.0.83
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.9.0
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pdf2image==1.17.0
pdfminer.six==20231228
pillow==10.2.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pypdf==4.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-slugify==8.0.2
python-webpack-boilerplate==1.0.2
pytz==2024.1
PyYAML==6.0.1
rcssmin==1.0.6
requests==2.31.0
rich==13.7.0
rjsmin==1.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
tesseract==0.1.3
text-unidecode==1.3
tqdm==4.66.1
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Wand==0.6.10
yarl==1.9.4
